refactor project structure : put all uninformed search algorithm in algorithm directory
